Render areas using gluTess(), all is always recalculated thou.

+++ OpenNavigator/GLNodes.h
@@ -13,21 +13,21 @@
 @private
     NSDictionary* _nodes;
     NSMutableDictionary* _indexes;
-    GLfloat* _nodeVertices;
+    GLdouble* _nodeVertices;
     double min_lat;
     double max_lat;
     double min_lon;
     double max_lon;
     size_t _bufferSize;
 }
-@property (readonly) GLfloat* nodeVertices;
+@property (readonly) GLdouble* nodeVertices;
 @property (readonly) size_t bufferSize;
 
 
 - (id) initWithNodes:(NSDictionary*)nodes andBounds:(NSRect)rect;
 - (NSUInteger)count;
-- (GLfloat)lonToLocal: (double)lon;
-- (GLfloat)latToLocal: (double)lat;
+- (GLdouble)lonToLocal: (double)lon;
+- (GLdouble)latToLocal: (double)lat;
 - (GLushort) indexOfNodeWithId: (NSNumber*)nodeid;
